+++ src/main/java/insane96mcp/iguanatweaksreborn/data/criterion/ITRTriggers.java
@@ -7,6 +7,7 @@ public class ITRTriggers {
     public static AnvilRepairTrigger ANVIL_REPAIR = CriteriaTriggers.register(new AnvilRepairTrigger());
     public static MakeRichFarmlandTrigger MAKE_RICH_FARMLAND = CriteriaTriggers.register(new MakeRichFarmlandTrigger());
     public static UnfairOneShotTrigger UNFAIR_ONESHOT = CriteriaTriggers.register(new UnfairOneShotTrigger());
+    public static SeasonChangedTrigger SEASON_CHANGED = CriteriaTriggers.register(new SeasonChangedTrigger());
 
     public static void init() {
         //Do nothing, it's just to trigger the registration of CriteriaTriggers